+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
--  `Connection.authenticate_oidc()`: add argument to set maximum device code flow poll time
+- `Connection.authenticate_oidc()`: add argument to set maximum device code flow poll time
+- Show progress bar while waiting for OIDC authentication with device code flow,
+  including special mode for in Jupyter notebooks.
+  ([#237](https://github.com/Open-EO/openeo-python-client/issues/237))
 
 ### Changed